Read TOTAL_POROSITY field if necessary
Up to now, total porosity has been calculated from column and particle
porosity. If a TOTAL_POROSITY field is present, the total porosity will
be read from it and used. However, if only COL_POROSITY and
PAR_POROSITY are present, they will be read and used for computing the
total porosity.

This change maintains compatibility to the CADET v3 file format.